Move components, change prefab
