+++ src/main.cpp
@@ -4,15 +4,12 @@
 #include "modules/main_menu.h"
 #include "modules/render.h"
 #include "modules/rocket_launch.h"
+#include "modules/site.h"
 #include "spdlog/sinks/basic_file_sink.h"
 #include "spdlog/spdlog.h"
 #include "systems.h"
-#include "systems/building_systems.h"
 
 void registerResources(flecs::world &world) {
-  world.set<PrefabResource>({
-      world.prefab("Rocket").add<Rocket>().set<CargoHold>({1000}), // Rocket
-  });
   // Find all Persons that are TeamMember of a $team, where the $team is a
   // TeamMember of anything
   // i.e. Match any person with a team that has another team above it.
@@ -72,11 +69,6 @@ void registerSystems(flecs::world &world) {
       .kind(UpdatePhase)
       .each(systemUpdateSimDate);
 
-  world.system<Manufacturing>("Update Construction")
-      .tick_source(game->sim_speed)
-      .kind(UpdatePhase)
-      .each(systemBuildingUpdateConstruction);
-
   world.system<GuiResource>("Draw Site Window")
       .term_at(0)
       .singleton()
@@ -114,9 +106,10 @@ int main(void) {
   registerSystems(world);
 
   world.import <RenderModule>();
-  world.import <MainMenuModule>();
-  world.import <RocketLaunchModule>();
   world.import <GuiModule>();
+  world.import <MainMenuModule>();
+  world.import <SiteModule>();
+  world.import <RocketLaunchModule>();
 
   auto site =
       world.entity("cape_canaveral").set<Site>({"Cape Canaveral", 10, 10});
