--- src/systems/building_systems.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "components.h"
-
-void systemBuildingUpdateConstruction(flecs::entity entity, Manufacturing &);
